+++ scripts/tools/note.js
@@ -65,14 +65,21 @@ function findEffectiveBPM(notes, bpm) {
         1: null,
         3: null,
     };
+    const swingNoteArray = {
+        0: [],
+        1: [],
+        3: [],
+    };
     for (let i = 0, len = notes.length; i < len; i++) {
         const note = notes[i];
         if (isNote(note) && lastNote[note._type]) {
-            if (swingNext(note, lastNote[note._type])) {
+            if (swingNext(note, lastNote[note._type], swingNoteArray[note._type])) {
                 EBPM = Math.max(EBPM, bpm / ((note._time - lastNote[note._type]._time) * 2));
+                swingNoteArray[note._type] = [];
             }
         }
         lastNote[note._type] = note;
+        swingNoteArray[note._type].push(note);
     }
     return EBPM;
 }
@@ -83,16 +90,23 @@ function findEffectiveBPMSwing(notes, bpm) {
         1: null,
         3: null,
     };
+    const swingNoteArray = {
+        0: [],
+        1: [],
+        3: [],
+    };
     for (let i = 0, len = notes.length; i < len; i++) {
         const note = notes[i];
         if (isNote(note) && lastNote[note._type]) {
-            if (swingNext(note, lastNote[note._type])) {
+            if (swingNext(note, lastNote[note._type], swingNoteArray[note._type])) {
                 EBPM = Math.max(EBPM, bpm / ((note._time - lastNote[note._type]._time) * 2));
                 lastNote[note._type] = note;
+                swingNoteArray[note._type] = [];
             }
         } else {
             lastNote[note._type] = note;
         }
+        swingNoteArray[note._type].push(note);
     }
     return EBPM;
 }
@@ -105,18 +119,25 @@ function getEffectiveBPMTime(diff, mapSettings) {
         1: null,
         3: null,
     };
+    const swingNoteArray = {
+        0: [],
+        1: [],
+        3: [],
+    };
     for (let i = 0, len = notes.length; i < len; i++) {
         const note = notes[i];
         let EBPM = 0;
         if (isNote(note) && lastNote[note._type]) {
-            if (swingNext(note, lastNote[note._type])) {
+            if (swingNext(note, lastNote[note._type], swingNoteArray[note._type])) {
                 EBPM = bpm / ((note._time - lastNote[note._type]._time) * 2);
+                swingNoteArray[note._type] = [];
             }
         }
-        lastNote[note._type] = note;
         if (EBPM > tool.ebpm.th) {
             arr.push(adjustTime(note._time, bpm, offset, bpmc));
         }
+        lastNote[note._type] = note;
+        swingNoteArray[note._type].push(note);
     }
     return arr;
 }
@@ -129,13 +150,19 @@ function getEffectiveBPMSwingTime(diff, mapSettings) {
         1: null,
         3: null,
     };
+    const swingNoteArray = {
+        0: [],
+        1: [],
+        3: [],
+    };
     for (let i = 0, len = notes.length; i < len; i++) {
         const note = notes[i];
         let EBPM = 0;
         if (isNote(note) && lastNote[note._type]) {
-            if (swingNext(note, lastNote[note._type])) {
+            if (swingNext(note, lastNote[note._type], swingNoteArray[note._type])) {
                 EBPM = bpm / ((note._time - lastNote[note._type]._time) * 2);
                 lastNote[note._type] = note;
+                swingNoteArray[note._type] = [];
             }
         } else {
             lastNote[note._type] = note;
@@ -143,6 +170,7 @@ function getEffectiveBPMSwingTime(diff, mapSettings) {
         if (EBPM > tool.ebpm.thSwing) {
             arr.push(adjustTime(note._time, bpm, offset, bpmc));
         }
+        swingNoteArray[note._type].push(note);
     }
     return arr;
 }
@@ -157,20 +185,25 @@ function getMinSliderSpeed(notes) {
         1: null,
         3: null,
     };
+    const swingNoteArray = {
+        0: [],
+        1: [],
+        3: [],
+    };
     for (let i = 0, len = notes.length; i < len; i++) {
         const note = notes[i];
         if (isNote(note) && lastNote[note._type]) {
-            if (!swingNext(note, lastNote[note._type])) {
+            if (!swingNext(note, lastNote[note._type], swingNoteArray[note._type])) {
                 sliderSpeed[note._type] = Math.max(
                     sliderSpeed[note._type],
                     toRealTime(note._time - lastNote[note._type]._time) / (swingWindow(note, lastNote[note._type]) ? 2 : 1)
                 );
+            } else {
+                swingNoteArray[note._type] = [];
             }
-            lastNote[note._type] = note;
-        } else {
-            lastNote[note._type] = note;
         }
         lastNote[note._type] = note;
+        swingNoteArray[note._type].push(note);
     }
     return Math.max(sliderSpeed[0], sliderSpeed[1]);
 }
@@ -185,19 +218,28 @@ function getMaxSliderSpeed(notes) {
         1: null,
         3: null,
     };
+    const swingNoteArray = {
+        0: [],
+        1: [],
+        3: [],
+    };
     for (let i = 0, len = notes.length; i < len; i++) {
         const note = notes[i];
         if (isNote(note) && lastNote[note._type]) {
-            if (!swingNext(note, lastNote[note._type]) && toRealTime(note._time - lastNote[note._type]._time) > 0.001) {
+            if (
+                !swingNext(note, lastNote[note._type], swingNoteArray[note._type]) &&
+                toRealTime(note._time - lastNote[note._type]._time) > 0.001
+            ) {
                 sliderSpeed[note._type] = Math.min(
                     sliderSpeed[note._type],
                     toRealTime(note._time - lastNote[note._type]._time) / (swingWindow(note, lastNote[note._type]) ? 2 : 1)
                 );
+            } else {
+                swingNoteArray[note._type] = [];
             }
-            lastNote[note._type] = note;
-        } else {
-            lastNote[note._type] = note;
         }
+        lastNote[note._type] = note;
+        swingNoteArray[note._type].push(note);
     }
     return Math.min(sliderSpeed[0], sliderSpeed[1]);
 }
@@ -221,10 +263,15 @@ function detectDoubleDirectional(diff, mapSettings) {
         1: null,
         3: null,
     };
+    const swingNoteArray = {
+        0: [],
+        1: [],
+        3: [],
+    };
     for (let i = 0, len = notes.length; i < len; i++) {
         const note = notes[i];
         if (isNote(note) && lastNote[note._type]) {
-            if (swingNext(note, lastNote[note._type])) {
+            if (swingNext(note, lastNote[note._type], swingNoteArray[note._type])) {
                 if (startNoteDot[note._type]) {
                     startNoteDot[note._type] = null;
                     lastNoteDirection[note._type] = flipCutDir[lastNoteDirection[note._type]];
@@ -237,6 +284,7 @@ function detectDoubleDirectional(diff, mapSettings) {
                 } else {
                     lastNoteDirection[note._type] = note._cutDirection;
                 }
+                swingNoteArray[note._type] = [];
             } else {
                 if (startNoteDot[note._type] && checkAngle(note._cutDirection, lastNoteDirection[note._type], 45)) {
                     arr.push(adjustTime(startNoteDot[note._type]._time, bpm, offset, bpmc));
@@ -252,6 +300,7 @@ function detectDoubleDirectional(diff, mapSettings) {
             lastNoteDirection[note._type] = note._cutDirection;
         }
         lastNote[note._type] = note;
+        swingNoteArray[note._type].push(note);
         if (note._type === 3) {
             // on bottom row
             if (note._lineLayer === 0) {
@@ -361,15 +410,21 @@ function detectOffPrecision(diff, mapSettings) {
         1: null,
         3: null,
     };
+    const swingNoteArray = {
+        0: [],
+        1: [],
+        3: [],
+    };
     for (let i = 0, len = notes.length; i < len; i++) {
         const note = notes[i];
         const noteTime = adjustTime(note._time, bpm, offset, bpmc);
         if (isNote(note) && lastNote[note._type]) {
-            if (swingNext(note, lastNote[note._type])) {
+            if (swingNext(note, lastNote[note._type], swingNoteArray[note._type])) {
                 if (checkPrec(noteTime)) {
                     arr.push(noteTime);
                 }
                 lastNote[note._type] = note;
+                swingNoteArray[note._type] = [];
             }
         } else {
             if (isNote(note) && checkPrec(noteTime)) {
@@ -377,6 +432,7 @@ function detectOffPrecision(diff, mapSettings) {
             }
             lastNote[note._type] = note;
         }
+        swingNoteArray[note._type].push(note);
     }
     return arr.filter(function (x, i, ary) {
         return !i || x !== ary[i - 1];
@@ -415,6 +471,11 @@ function detectHitboxStaircase(diff, mapSettings) {
         1: null,
         3: null,
     };
+    const swingNoteArray = {
+        0: [],
+        1: [],
+        3: [],
+    };
     const noteOccupy = {
         0: { _lineIndex: 0, _lineLayer: 0 },
         1: { _lineIndex: 0, _lineLayer: 0 },
@@ -423,7 +484,7 @@ function detectHitboxStaircase(diff, mapSettings) {
     for (let i = 0, len = notes.length; i < len; i++) {
         const note = notes[i];
         if (isNote(note) && lastNote[note._type]) {
-            if (swingNext(note, lastNote[note._type])) {
+            if (swingNext(note, lastNote[note._type], swingNoteArray[note._type])) {
                 lastSpeed[note._type] = toRealTime(note._time - lastNote[note._type]._time);
                 if (note._cutDirection !== 8) {
                     noteOccupy[note._type]._lineIndex = note._lineIndex + swingCutDirectionSpace[note._cutDirection][0];
@@ -432,6 +493,7 @@ function detectHitboxStaircase(diff, mapSettings) {
                     noteOccupy[note._type]._lineIndex = -1;
                     noteOccupy[note._type]._lineLayer = -1;
                 }
+                swingNoteArray[note._type] = [];
                 lastNoteDirection[note._type] = note._cutDirection;
             } else if (swingNoteEnd(note, lastNote[note._type], lastNoteDirection[note._type])) {
                 if (note._cutDirection !== 8) {
@@ -472,6 +534,7 @@ function detectHitboxStaircase(diff, mapSettings) {
             lastNoteDirection[note._type] = note._cutDirection;
         }
         lastNote[note._type] = note;
+        swingNoteArray[note._type].push(note);
     }
     return arr;
 }
@@ -495,10 +558,15 @@ function detectShrAngle(diff, mapSettings) {
         1: null,
         3: null,
     };
+    const swingNoteArray = {
+        0: [],
+        1: [],
+        3: [],
+    };
     for (let i = 0, len = notes.length; i < len; i++) {
         const note = notes[i];
         if (isNote(note) && lastNote[note._type]) {
-            if (swingNext(note, lastNote[note._type])) {
+            if (swingNext(note, lastNote[note._type], swingNoteArray[note._type])) {
                 if (startNoteDot[note._type]) {
                     startNoteDot[note._type] = null;
                     lastNoteDirection[note._type] = flipCutDir[lastNoteDirection[note._type]];
@@ -514,6 +582,7 @@ function detectShrAngle(diff, mapSettings) {
                 } else {
                     lastNoteDirection[note._type] = note._cutDirection;
                 }
+                swingNoteArray[note._type] = [];
             } else {
                 if (
                     startNoteDot[note._type] &&
@@ -531,6 +600,7 @@ function detectShrAngle(diff, mapSettings) {
             lastNoteDirection[note._type] = note._cutDirection;
         }
         lastNote[note._type] = note;
+        swingNoteArray[note._type].push(note);
     }
     return arr.filter(function (x, i, ary) {
         return !i || x !== ary[i - 1];
@@ -615,10 +685,15 @@ function detectSpeedPause(diff, mapSettings) {
         1: false,
         3: false,
     };
+    const swingNoteArray = {
+        0: [],
+        1: [],
+        3: [],
+    };
     for (let i = 0, len = notes.length; i < len; i++) {
         const note = notes[i];
         if (isNote(note) && lastNote[note._type]) {
-            if (swingNext(note, lastNote[note._type])) {
+            if (swingNext(note, lastNote[note._type], swingNoteArray[note._type])) {
                 if (isBelowThres(note._time - lastNote[note._type]._time, tool.maxSpeedPause * 2 + 0.01)) {
                     if (
                         maybePause[0] &&
@@ -635,11 +710,13 @@ function detectSpeedPause(diff, mapSettings) {
                     maybePause[note._type] = true;
                     lastNotePause[note._type] = lastNote[note._type];
                 }
+                swingNoteArray[note._type] = [];
                 lastNote[note._type] = note;
             }
         } else {
             lastNote[note._type] = note;
         }
+        swingNoteArray[note._type].push(note);
     }
     return arr;
 }
@@ -669,11 +746,17 @@ function detectImproperWindowSnap(diff, mapSettings) {
         1: null,
         3: null,
     };
+    const swingNoteArray = {
+        0: [],
+        1: [],
+        3: [],
+    };
     for (let i = 0, len = notes.length; i < len; i++) {
         const note = notes[i];
         if (isNote(note) && lastNote[note._type]) {
-            if (swingNext(note, lastNote[note._type])) {
+            if (swingNext(note, lastNote[note._type], swingNoteArray[note._type])) {
                 lastNote[note._type] = note;
+                swingNoteArray[note._type] = [];
             } else if (
                 isSlantedWindow(note, lastNote[note._type]) &&
                 isAboveThres(note._time - lastNote[note._type]._time, tool.windowSnapTolerance) &&
@@ -686,6 +769,7 @@ function detectImproperWindowSnap(diff, mapSettings) {
         } else {
             lastNote[note._type] = note;
         }
+        swingNoteArray[note._type].push(note);
     }
     return arr.filter(function (x, i, ary) {
         return !i || x !== ary[i - 1];
@@ -714,26 +798,32 @@ function detectSlowSlider(diff, mapSettings) {
         1: 0,
         3: 0,
     };
+    const swingNoteArray = {
+        0: [],
+        1: [],
+        3: [],
+    };
     for (let i = 0, len = notes.length; i < len; i++) {
         const note = notes[i];
         if (isNote(note) && lastNote[note._type]) {
-            if (swingNext(note, lastNote[note._type])) {
+            if (swingNext(note, lastNote[note._type], swingNoteArray[note._type])) {
                 sliderSpeed[note._type] = 0;
                 lastNoteTime[note._type] = note._time;
+                swingNoteArray[note._type] = [];
             } else {
                 sliderSpeed[note._type] = Math.max(
                     sliderSpeed[note._type],
                     toRealTime(note._time - lastNote[note._type]._time) / (swingWindow(note, lastNote[note._type]) ? 2 : 1)
                 );
             }
-            lastNote[note._type] = note;
             if (sliderSpeed[note._type] > tool.minSliderSpeed) {
                 arr.push(adjustTime(lastNoteTime[note._type], bpm, offset, bpmc));
             }
         } else {
             lastNoteTime[note._type] = note._time;
-            lastNote[note._type] = note;
         }
+        lastNote[note._type] = note;
+        swingNoteArray[note._type].push(note);
     }
     return arr.filter(function (x, i, ary) {
         return !i || x !== ary[i - 1];
@@ -760,12 +850,17 @@ function detectInlineAngle(diff, mapSettings) {
         1: null,
         3: null,
     };
+    const swingNoteArray = {
+        0: [],
+        1: [],
+        3: [],
+    };
     let lastTime = 0;
     let lastIndex = 0;
     for (let i = 0, len = notes.length; i < len; i++) {
         const note = notes[i];
         if (isNote(note) && lastNote[note._type]) {
-            if (swingNext(note, lastNote[note._type])) {
+            if (swingNext(note, lastNote[note._type], swingNoteArray[note._type])) {
                 if (startNoteDot[note._type]) {
                     startNoteDot[note._type] = null;
                     lastNoteDirection[note._type] = flipCutDir[lastNoteDirection[note._type]];
@@ -778,6 +873,7 @@ function detectInlineAngle(diff, mapSettings) {
                 } else {
                     lastNoteDirection[note._type] = note._cutDirection;
                 }
+                swingNoteArray[note._type] = [];
             } else {
                 if (
                     startNoteDot[note._type] &&
@@ -794,11 +890,12 @@ function detectInlineAngle(diff, mapSettings) {
         } else {
             lastNoteDirection[note._type] = note._cutDirection;
         }
-        lastNote[note._type] = note;
         if (lastTime < note._time - toBeatTime(tool.maxInlineAngle - 0.01)) {
             lastTime = note._time - toBeatTime(tool.maxInlineAngle - 0.01);
             lastIndex = i;
         }
+        lastNote[note._type] = note;
+        swingNoteArray[note._type].push(note);
         if (note._type === 3) {
             // on bottom row
             if (note._lineLayer === 0) {
@@ -844,12 +941,95 @@ function checkInline(note, notes, index) {
     }
     return false;
 }
-function checkAngle(n1cd, n2cd, angle) {
+function checkAngle(n1cd, n2cd, angle, invert = false) {
     if (n1cd === 8 || n2cd === 8) {
         return false;
     }
-    if (shortRotDistance(noteCutAngle[n1cd], noteCutAngle[n2cd], 360) <= angle) {
+    if (
+        !invert
+            ? shortRotDistance(noteCutAngle[n1cd], noteCutAngle[n2cd], 360) <= angle
+            : shortRotDistance(noteCutAngle[n1cd], noteCutAngle[n2cd], 360) >= angle
+    ) {
         return true;
     }
     return false;
+}
+
+function detectReverseStaircase(diff, mapSettings) {
+    const { _notes: notes } = diff;
+    const { njs, bpm, bpmc, offset } = mapSettings;
+    const arr = [];
+    const lastNote = {
+        0: null,
+        1: null,
+        3: null,
+    };
+    const swingNoteArray = {
+        0: [],
+        1: [],
+        3: [],
+    };
+    for (let i = 0, len = notes.length; i < len; i++) {
+        const note = notes[i];
+        if (isNote(note) && lastNote[note._type]) {
+            if (swingNext(note, lastNote[note._type], swingNoteArray[note._type])) {
+                swingNoteArray[note._type] = [];
+            }
+        }
+        for (const other of swingNoteArray[(note._type + 1) % 2]) {
+            if (other._cutDirection !== 8) {
+                let noteOccupyLineIndex = other._lineIndex + swingCutDirectionSpace[flipCutDir[other._cutDirection]][0];
+                let noteOccupyLineLayer = other._lineLayer + swingCutDirectionSpace[flipCutDir[other._cutDirection]][1];
+                if (
+                    !(njs > bpm / (120 * (note._time - other._time))) &&
+                    note._lineIndex === noteOccupyLineIndex &&
+                    note._lineLayer === noteOccupyLineLayer
+                ) {
+                    arr.push(adjustTime(other._time, bpm, offset, bpmc));
+                    break;
+                }
+            }
+        }
+        lastNote[note._type] = note;
+        swingNoteArray[note._type].push(note);
+    }
+    return arr;
+}
+
+function detectInlineHitbox(diff, mapSettings) {
+    const { _notes: notes } = diff;
+    const { njs, bpm, bpmc, offset } = mapSettings;
+    const arr = [];
+    const lastNote = {
+        0: null,
+        1: null,
+        3: null,
+    };
+    const swingNoteArray = {
+        0: [],
+        1: [],
+        3: [],
+    };
+    for (let i = 0, len = notes.length; i < len; i++) {
+        const note = notes[i];
+        if (isNote(note) && lastNote[note._type]) {
+            if (swingNext(note, lastNote[note._type], swingNoteArray[note._type])) {
+                swingNoteArray[note._type] = [];
+            }
+        }
+        for (const other of swingNoteArray[(note._type + 1) % 2]) {
+            // magic number 1.425 from saber length + good/bad hitbox
+            if (
+                !(njs > (1.425 * bpm) / (120 * (note._time - other._time))) &&
+                note._lineIndex === other._lineIndex &&
+                note._lineLayer === other._lineLayer
+            ) {
+                arr.push(adjustTime(note._time, bpm, offset, bpmc));
+                break;
+            }
+        }
+        lastNote[note._type] = note;
+        swingNoteArray[note._type].push(note);
+    }
+    return arr;
 }